requirements.txt
-e hg+https://bitbucket.org/pyglet/pyglet#egg=pyglet

cairocffi
nanomsg
msgpack-python
